backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
httpx==0.25.2
apscheduler==3.10.4
openai==1.3.7
beautifulsoup4==4.12.2
lxml==4.9.3
aiofiles==23.2.1
alembic==1.13.0
python-dotenv==1.0.0